Moving FontMenu and Heading files to the separate folder
